+++ src/main/java/com/clova/anifriends/domain/review/Review.java
@@ -5,7 +5,7 @@ import com.clova.anifriends.domain.applicant.wrapper.ApplicantStatus;
 import com.clova.anifriends.domain.common.BaseTimeEntity;
 import com.clova.anifriends.domain.review.exception.ReviewAuthorizationException;
 import com.clova.anifriends.domain.review.exception.ReviewBadRequestException;
-import com.clova.anifriends.domain.review.wrapper.ReviewContent;
+import com.clova.anifriends.domain.review.vo.ReviewContent;
 import com.clova.anifriends.domain.volunteer.Volunteer;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
